Flytt hjelpefunksjoner for autentisering til egen fil
