+++ src/main/kotlin/rekrutteringsbistand/stilling/indekser/App.kt
@@ -1,14 +1,10 @@
 package rekrutteringsbistand.stilling.indekser
 
 import com.github.kittinunf.fuel.core.FuelManager
-import com.github.kittinunf.fuel.core.extensions.authentication
 import io.javalin.Javalin
 import io.javalin.apibuilder.ApiBuilder.get
-import io.javalin.http.NotFoundResponse
-import org.eclipse.jetty.http.HttpStatus
-import rekrutteringsbistand.stilling.indekser.autentisering.AccessToken
 import rekrutteringsbistand.stilling.indekser.autentisering.AccessTokenClient
-import rekrutteringsbistand.stilling.indekser.stillingsinfo.StillingsinfoClient
+import rekrutteringsbistand.stilling.indekser.autentisering.authenticateAllRequests
 
 class App {
     companion object {
@@ -32,22 +28,3 @@ fun main() {
     App.start(authenticatedClient)
 }
 
-fun authenticateAllRequests(httpClient: FuelManager, accessTokenClient: AccessTokenClient): FuelManager {
-    val rekrutteringsbistandApiClientId = "fe698176-ac44-4260-b8d0-dbf45dd956cf"
-
-    addAccessToken(httpClient) {
-        accessTokenClient.getAccessToken(scope = "api://$rekrutteringsbistandApiClientId/.default")
-    }
-
-    return httpClient
-}
-
-fun addAccessToken(httpClient: FuelManager, getToken: () -> AccessToken) {
-    httpClient.addRequestInterceptor {
-        { request ->
-            val token = getToken()
-            request.authentication().bearer(token.access_token)
-            it(request)
-        }
-    }
-}
